requirements.txt
git+https://github.com/pyrogram/pyrogram
filesplit
wget>=3.2
CairoSVG>=2.5.2
faker>=6.6.2
html5lib
langdetect>=1.0.8
geopy>=2.1.0
PyYAML>=5.4.1
selenium>=3.141.0
opencv-contrib-python>=4.5.1.48
fake_useragent>=0.1.11
scipy>=1.6.1
numpy>=1.20.1
NoteShrinker>=0.2.0
pytest>=6.2.2
glitch_this>=1.0.2
lottie>=0.6.6
pygifsicle>=1.0.2
psutil>=5.8.0
python-dotenv>=0.15.0
pymediainfo>=5.0.3
motor>=2.3.1
youtube-search-python>=1.4.2
dnspython>=2.1.0
Markdown>=3.3.4
gTTS>=2.2.2
youtube-search>=2.1.0
ffmpeg-python>=0.2.0
wheel>=0.36.2
anime_downloader>=5.0.7
hachoir>=3.1.2
google-trans-new>=1.1.9
googletrans>=3.0.0
beautifulsoup4>=4.9.3
tinydb>=4.4.0
img2pdf>=0.4.0
gitpython>=3.1.14
TgCrypto>=1.2.2
pytz>=2021.1
apscheduler>=3.7.0
requests>=2.25.1
git+https://github.com/ytdl-org/youtube-dl
Pillow
pyshorteners>=1.0.1
telegraph>=1.4.1
heroku3>=4.2.3
aiofiles>=0.6.0
aiohttp>=3.7.4.post0
